feat: dfs router
